requirements.txt
absl-py==2.1.0
agora-token-builder==1.0.0
altair==5.2.0
annotated-types==0.7.0
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.1.0
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0
docutils==0.21.2
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.1
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozendict==2.4.0
fsspec==2024.10.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h5py==3.10.0
html5lib==1.1
httplib2==0.22.0
idna==3.6
image==1.5.33
importlib-metadata==7.0.1
inflection==0.5.1
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.3
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.15.0
keyring==25.6.0
kiwisolver==1.4.4
libclang==16.0.6
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplot==0.1.9
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==10.2.0
mpmath==1.3.0
multitasking==0.0.11
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nh3==0.2.20
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.25.2
oauthlib==3.2.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
peewee==3.17.0
pgmpy==0.1.26
Pillow==10.0.0
pkginfo==1.12.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.25.0
protobuf==4.23.4
psutil==5.9.7
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.8.1b0
pygame==2.5.2
Pygments==2.17.2
pyloco==0.0.139
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
Quandl==3.7.0
readme_renderer==44.0
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
Send2Trash==1.8.2
SimpleWebSocketServer==0.1.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.30.0
sympy==1.13.1
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.5.1
tornado==6.4
tqdm==4.66.6
traitlets==5.14.1
twine==6.0.1
types-python-dateutil==2.8.19.20240106
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
ushlex==0.99.1
validators==0.22.0
vboxapi==1.0
virtualenv==20.25.0
watchdog==3.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
xgboost==2.1.2
yfinance==0.2.35
zipp==3.17.0
